Setup MarkovAndMe Application
This is the base application for MarkovAndMe... Which is a text
generation server based on Markov Chains. It's trained by a json file
containing A list of list of Strings. Basically a 2d matrix of strings.
And it goes through each list of words, building up a chain of
probabilities between each word. Once that's completed, it serves the
results on a web server. This can easily be forked and modified, and
can be used for a variety of text generation needs.
The repo is also setup for the application to run on IBM's bluemix. so
no alterations should be needed to do that. although ... the
docker-compose file does not currently work.